Fix almost all test issues
Had to switch to LXD instead of MicroK8s because of the way
layer:caas-base builds the "venv" for K8s charms. Specifically, it drops
all package metadata so entry points don't work at all. It's also
missing all of the improvements to the composite wheelhouse.txt
processing that were added to charms.reactive. Will need to fix those
things in that layer.